% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afforestation.r
\name{afforestation}
\alias{afforestation}
\title{Afforestation}
\usage{
afforestation(land, clim, params)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records and land-cover types in rows}

\item{clim}{A data frame with minimum temperature (in ºC), maximum temperature (in ºC), 
and precipitation (in mm) per location}

\item{params}{A list of model parameters, default ones are generated by the function \code{default.params()}}
}
\value{
A vector with the new tree species in colonized shurb locations
}
\description{
Determines the shrub locations colonized by tree species
}
\examples{
data(landscape)
data(clim)
params = default.params()
afforestation(landscape, clim, params)

}
