% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{clim}
\alias{clim}
\title{Reference climate}
\format{
A data frame with 3210899 rows and 4 variables:
\describe{
  \item{cell.id}{Unique grid cell indentificator}
  \item{tmin}{Historical (1971-2000) mean minimum temperature, in ºC}
  \item{tmax}{Historical (1971-2000) mean maximum temperature, in ºC}
  \item{precip}{Historical (1971-2000) accumulated precipitation, in mm}
}
}
\usage{
clim
}
\description{
Climatic variables of the reference period 1971 - 2000
}
\keyword{datasets}
