% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{coord}
\alias{coord}
\title{UTM coordinates}
\format{
A data frame with 3210899 rows and 3 variables:
\describe{
  \item{cell.id}{Unique grid cell indentificator}
  \item{x}{UTM coordinate X}
  \item{y}{UTM coordinate Y}
}
}
\usage{
coord
}
\description{
Coordinates x and y in UTM31N-ETRS89 of the 1 ha cells in the study area.
}
\keyword{datasets}
