% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.params.r
\name{default.params}
\alias{default.params}
\title{Default model parameters}
\usage{
default.params()
}
\value{
A list with the following items:
 \itemize{
   \item{\code{species}: Vector with the name of the tree species (sorted as codified in the \code{landscape} data frame)}
   \item{\code{spp.distrib.rad}: Radius of the neighborhood (in pixels) to find out if a species is present in a region}
   \item{\code{colon.rad}: Radius of the neighborhood (in pixels) to look for mature tree species ready to colonize shrubs or locations killed by drought}
   \item{\code{rpb}: Spread rate}
   \item{\code{pb.upper.th}: If spread rate is above this threshold, the cell always burns}
   \item{\code{pb.lower.th}: If spread rate is below this threshold, the cell never burns}
   \item{\code{fire.intens.th}: high vs. low intensity fire, sr <= fire.intens.th}
   \item{\code{pb.target.area}: if NA, then burnt as \code{7*pb.convenient.area}, otherwise annually burnt pb.fix.area}
   \item{\code{pb.convenient.area}}
   \item{\code{accum.burnt.area}}
   \item{\code{pb.mean}}
   \item{\code{pb.sd}: Standard deviation of Normal distribution to drawn prescribed burn sizes}
   \item{\code{pb.fage.th}: Minimum forest age to apply prescribed burns, in years}
 }
}
\description{
Returns a list with the Quebec Landscape Dynamic Model parameters and the default values
}
\examples{
# Change the convenient area to be burnt by prescribed burns to 15.000 ha 
params = default.params()
params$pb.convenient.area = 15000

}
