% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drought.r
\name{drought}
\alias{drought}
\title{Drought mortality}
\usage{
drought(land, decade, t)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records and land-cover types in rows}

\item{decade}{A number indicating the decade of the current time step}

\item{t}{The time step}

\item{clim}{A data frame with minimum temperature (in ºC), maximum temperature (in ºC), and precipitation (in mm) per location}

\item{params}{A list of default parameters generated by the function \code{default.params()} or}
}
\value{
A vector with the \code{cell.id} of the killed locations
}
\description{
Determines the dieback of forest areas due to drought-induced mortatlidy
}
\examples{
data(landscape)
data(clim)
drought(landscape, clim, 10, 1)

}
