% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encroachment.r
\name{encroachment}
\alias{encroachment}
\title{Encroachment}
\usage{
encroachment(land)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records and land-cover types in rows}
}
\value{
A vector with the \code{cell.id} of the colonized grassland areas
}
\description{
Determines the grassland locations colonized by shrubs
}
\examples{
data(landscape)
encroachment(landscape)

}
