% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.regime.r
\name{fire.regime}
\alias{fire.regime}
\title{Fire regime}
\usage{
fire.regime(
  land,
  clim,
  params,
  swc = 1,
  clim.sever = 0,
  annual.burnt.area = 0,
  step = 1,
  out.path
)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records and land-cover types in rows}

\item{clim}{A data frame with minimum temperature (in ºC), maximum temperature (in ºC), 
and precipitation (in mm) per location}

\item{params}{A list of model parameters, default ones are generated by the function \code{default.params()}}

\item{swc}{Vector with the numeric indicator of the synoptic weather conditions to simulate: 
1 - wind, 2 - heat, 3 - regular, or 4 - prescribed burns}

\item{clim.sever}{Climatic severity of the year: 0 - mild, 1 - severe}

\item{annual.burnt.area}{The total burnt area in the current time stepA string indicating which regional policy is adopted to allocate the timber demands}

\item{step}{Integer indicating the current time step}

\item{out.path}{String with the directory path to save log files}
}
\value{
A list of two objects: A data frame with target, burnt and suppressed areas per fire, and 
a data frame with fire identificator, step of spreading, fire intensity and whether it is ignition (TRUE or FALSE) 
per burnt location (identified by \code{cell.id})
}
\description{
Wildfires under a synoptic weather condition or prescribed burns
}
\examples{
data(landscape)
data(clim)
params = default.params()
land = landscape
land$interface = interface(landscape)
# Simulate wildfires under Wind synoptic weather conditions for a climatic mild year
fire.regime(land, clim, params, swc = 1, clim.sever = 0, annual.burnt.area = 0, step = 1, out.path=tempdir())

}
