% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.r
\name{growth}
\alias{growth}
\title{Vegetation growth}
\usage{
growth(land, clim, who = "")
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records and land-cover types in rows}

\item{clim}{A data frame with minimum temperature (in ºC), maximum temperature (in ºC), 
and precipitation (in mm) per location}

\item{who}{String to display to what type of communitites is growing}
}
\value{
A vector with the new basal area or biomass for forest stands or shrublands, respectively
}
\description{
Growht of basal area of forest stands and biomass of shrublands
}
\examples{
data(landscape)
data(clim)
growth(landscape, clim)

}
