% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{harvest}
\alias{harvest}
\title{Timber harvesting allocation}
\format{
A data frame with 3210899 rows and 6 variables:
\describe{
  \item{cell.id}{Unique grid cell indentificator}
  \item{enpe}{Code of the protection status, 0 - none, 1-national park}
  \item{dist.path}{Distance to the nearest path or road, in m}
  \item{slope.pctg}{Stand slope, in /100}
  \item{dist.industry}{Distance to the nearest wood transformation industry, in km}
  \item{dist.biomass}{Distance to the nearest biomass plant, in km}
}
}
\usage{
harvest
}
\description{
Landscape variables influencing the allocation of sylvicultural prescriptions
}
\keyword{datasets}
