% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.r
\name{interface}
\alias{interface}
\title{Land interfaces}
\usage{
interface(land, orography)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records and land-cover types in rows}

\item{orography}{A data frame with the orographic characteristics of the study area}
}
\value{
A vector with the code interface of all locations in the study area
}
\description{
Determines the land interface of each location
}
\examples{
data(landscape)
data(orography)
afforestation(landscape, coord, orography, clim, sdm)

}
