% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.cover.change.r
\name{land.cover.change}
\alias{land.cover.change}
\title{Land-cover changes}
\usage{
land.cover.change(land, lc.trans, target.demand, visit.cells)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records and land-cover types in rows}

\item{lc.trans}{A number indicating the land-cover transition to simulate: 1 - urbanization, 2 - agriculture 
conversion, and 3 - rural abandonment}

\item{target.demand}{Area to be converted to the target land-cover (in ha)}

\item{visit.cells}{Vector with the \code{cell.id} of already changed cells in the current time step}
}
\value{
A vector with the \code{cell.id} of the colonized grassland areas
}
\description{
Spatially allocation of land-cover transitions
}
\examples{
data(landscape)
data(coord)
land = interface(landscape)
land.cover.change(land, 1, 10, numeric())

}
