% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.dyn.mdl.r
\name{land.dyn.mdl}
\alias{land.dyn.mdl}
\title{The medLDM}
\usage{
land.dyn.mdl(
  is.land.cover.change = FALSE,
  is.harvest = FALSE,
  is.wildfire = FALSE,
  is.prescribed.burn = FALSE,
  is.drought = TRUE,
  is.postfire = TRUE,
  is.cohort.establish = TRUE,
  is.afforestation = TRUE,
  is.encroachment = TRUE,
  is.growth = TRUE,
  spin.up = TRUE,
  custom.params = NA,
  clim.proj = NA,
  nrun = 1,
  time.step = 1,
  clim.step = 10,
  time.horizon = 90,
  save.land = FALSE,
  out.seq = NA,
  out.path = NA,
  ...
)
}
\arguments{
\item{is.land.cover.change}{A flag to indicate that land cover changes are simulated}

\item{is.harvest}{A flag to indicate that harvesting for sawlogs and wood is simulated}

\item{is.drought}{A flag to indicate that prescribed burns are simulated}

\item{is.postfire}{A flag to indicate that prescribed burns are simulated}

\item{is.cohort.establish}{A flag to indicate that prescribed burns are simulated}

\item{is.afforestation}{A flag to indicate that prescribed burns are simulated}

\item{is.encroachment}{A flag to indicate that prescribed burns are simulated}

\item{is.growth}{A flag to indicate that prescribed burns are simulated}

\item{spin.up}{A flag to indicate if the observed 2010-2019 wildfires and land-cover changes are replicated}

\item{custom.params}{List with the model paramaters and default and/or user-defined values}

\item{clim.proj}{A list of data frame with projections of climatic variables (minimum temperature, annual precipitation) for each \code{clim.step }}

\item{nrun}{Number of replicates to run the model}

\item{time.step}{Number of years of each climatic period}

\item{time.horizon}{Number of years of the model simulation, it has to be a multiple \code{time.step}}

\item{save.land}{A flag to save as a RDS file the \code{landscape} data frame at the time step indicated in \code{out.seq}}

\item{out.seq}{Numeric vector with the time steps the \code{landscape} is saved}

\item{out.path}{String with the directory path to save the \code{landscape} data frame at each time step indicated in \code{out.seq}}

\item{is.widlfire}{A flag to indicate that wildfires are simualted}

\item{is.prescribed.burnt}{A flag to indicate that prescribed burns are simulated}
}
\value{
A list with the following items:
 \itemize{
   \item{\code{SppByAgeClass}: A data frame of species abundance per age class, with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{mgmt.unit}: Code of the forest management unit (FMU).}
        \item{\code{spp}: Code of the species.}
        \item{\code{age.class}: Code of the age class, \code{C10}, \code{C30}, \code{C50}, \code{C70}, \code{C90}, and \code{OLD}.}
        \item{\code{area}: Area in \eqn{km^{2}}.}
      }
   }
   \item{\code{SuitabilityClass}: A data frame of suitability of potential species per bioclimatic domain, with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{bioclim.domain}: Code of the bioclimatic domain.}
        \item{\code{spp}: Code of the species.}
        \item{\code{poor}: Area in \eqn{km^{2}} of poor environmental suitability.}
        \item{\code{med}: Area in \eqn{km^{2}} of intermediate environmental suitability.}
        \item{\code{good}: Area in \eqn{km^{2}} of good environmental suitability.}
      }
   }
   \item{\code{SppByFireZone}: A data frame of species abundance per fire zone, with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{frz}: Code of the fire regime zone.}
        \item{\code{spp}: Code of the species.}
        \item{\code{area}: Area in \eqn{km^{2}}.}
      }
   }
   \item{\code{FuelByFireZone}: A data frame of fuel type per fire zone, with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{frz}: Code of the fire regime zone.}
        \item{\code{type}: Code of the fuel type:  \code{low}, \code{med} or \code{high}.}
        \item{\code{pct}: Relative abundance of the fuel type in the fire regime zone ([0,1]).}
      }
   }
   \item{\code{Cuts}: A data frame of harvestable area and volume per management unit 
   (included if \code{is.harvesting}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{mgmt.unit}: Code of the forest management unit (FMU).}
        \item{\code{a.inc}: Area that can be harvested, i.e. non-protected and FMU informed (in \eqn{km^{2}}).}
        \item{\code{a.even.age}: Area within \code{a.inc} of mature stand, i.e.\code{age>age.matu} (in \eqn{km^{2}}).}
        \item{\code{a.mat.pc}: Area to be managed under a partial-cut regime (in \eqn{km^{2}}).}
        \item{\code{a.inc.burnt}: Area within \code{a.inc} burnt in the current time step (in \eqn{km^{2}}).}
        \item{\code{a.inc.mat.burnt}: Area within \code{a.inc} of mature stands burnt in the current time step (in \eqn{km^{2}}).}
        \item{\code{a.inc.kill}: Area within \code{a.inc} killed by SBW in the previous or current time step (in \eqn{km^{2}}).}
        \item{\code{a.inc.mat.kill}: Area within \code{a.inc} of mature stands killed by SBW in the previous or current time step (in \eqn{km^{2}}).}
        \item{\code{a.reg.fail.ex}: Protected area at defforestation risk (in \eqn{km^{2}}).}
        \item{\code{a.reg.fail.in}: Non-protected area at defforestation risk (in \eqn{km^{2}}).}
        \item{\code{a.salvaged}: Salvaged and clear-cut area (in \eqn{km^{2}}).}
        \item{\code{a.unaff}: Clear-cut area unaffected by disturbances (in \eqn{km^{2}}).}
        \item{\code{v.salv}: Salvaged and clear-cut volume (in \eqn{m^{3}}).}
        \item{\code{v.unaff}: Clear-cut volume unaffected by disturbances (in \eqn{m^{3}}).}
        \item{\code{a.pcut}: Area partial cut (in \eqn{km^{2}}).}
        \item{\code{v.pcut}: Volume partial cut (in \eqn{m^{3}}).}
      }
   }
   \item{\code{SppCut}: A data frame of area and volum extracted by clear and partial cut per species 
   and management unit (included if \code{is.harvesting}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{mgmt.unit}: Code of the forest management unit (FMU).}
        \item{\code{spp}: Code of the species.}
        \item{\code{area.ccut}: Clear-cut area (in \eqn{km^{2}}).}
        \item{\code{vol.ccut}: Clear-cut volume (in \eqn{m^{3}}).}
        \item{\code{area.pcut}: Partial cut area (in \eqn{km^{2}}).}
        \item{\code{vol.pcut}: Partial cut volume (in \eqn{m^{3}}).}
      }
   }
   \item{\code{BurntRates}: A data frame of target area to be burnt per fire regime zone 
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{frz}: Code of the fire regime zone.}
        \item{\code{br}: Baseline area to be burnt derived from MFRI (in \eqn{km^{2}}).}
        \item{\code{brvar}: Baseline area \code{br} with inter-period variability added (in \eqn{km^{2}}).}
        \item{\code{brfuel}: Variable baseline area \code{brvar} modified according to zone-level fuel if
        \code{is.fuel.modifier} (in \eqn{km^{2}}).}
        \item{\code{brclima}: ariable baseline area \code{brvar} modified according to zone-level SEP rate if
        \code{is.clima.modifier}  (in \eqn{km^{2}}).}
        \item{\code{target.area}: Target area to be burnt (in \eqn{km^{2}}).}
      }
   }
   \item{\code{FireRegime}: A data frame of number of fires and burnt area per fire regime zone
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{frz}: Code of the fire regime zone.}
        \item{\code{target.area}: Target area to be burnt (in \eqn{km^{2}}).}
        \item{\code{nfires}: Number of fires burnt.}
        \item{\code{burnt.area}: Area effectively burnt (in \eqn{km^{2}}).}
        \item{\code{fire.cycle}: Relative fire return interval (in years).}
        \item{\code{indx.combust}: Fire-zone mean fuel flammability ([0,1]).}
        \item{\code{indx.combust.burnt}: Mean fuel flammability of actual burnt areas ([0,1]).}
      }
   }
   \item{\code{Fires}: A data frame of wildfires
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{frz}: Code of the fire regime zone.}
        \item{\code{fire.id}: Wildfire identificator.}
        \item{\code{wind}: Main wind direction in degrees.}
        \item{\code{target.size}: Target area to be burnt (in \eqn{km^{2}}).}
        \item{\code{burnt.size}: Area effectively burnt (in \eqn{km^{2}}).}
        \item{\code{rem}: Remanent area to be burnt (in \eqn{km^{2}}).}
      }
   }
   \item{\code{BurntFuels}: A data frame of burnt fuel types per fire regime zone
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{frz}: Code of the fire regime zone.}
        \item{\code{type}: Code of the fuel type:  \code{low}, \code{med} or \code{high}.}
        \item{\code{area}: Area burnt of each fuel category (in \eqn{km^{2}}).}
      }
   }
 }
}
\description{
Run the Mediterranean Landscape Dynamic Model medLDM that includes the processes of land-cover changes
wilfires, prescribed burns, fire suppression, timber and wood harvesting and vegetation dynamics.
}
\examples{

\dontrun{
library(medLDM)
# Run one single 80-year replicate with forest management and default RCP 4.5 climate projections 
ldm(is.harvest = T)
}

}
