% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{landscape}
\alias{landscape}
\title{Forest landscape features of Catalonia}
\format{
A data frame with 3210899 rows and 6 variables:
\describe{
  \item{cell.id}{Unique grid cell indentificator}
  \item{spp}{Tree species or land-cover types: 1 - Pinus halepensis, 2 - }
  \item{biom}{Unique grid cell indentificator}
  \item{age}{Forest and shrub age, in year}
  \item{typdist}{Type of the last disturbance, possible values are \code{highfire}, \code{lowfire}}
  \item{tsdist}{Time since the last disturbance, in year}
  \item{x}{Interface code: 1 - urban, 2 - cropland, 3 - natural areas, 4 - other, 
  5 - urban & cropland, 6 - urban & natural areas, and 7 - cropland & natural areas}
}
}
\usage{
landscape
}
\description{
Landscape and forest stands characteristics of Catalonia in 2020 at 1 ha of spatial resolution
}
\keyword{datasets}
