% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medLDM.r
\docType{package}
\name{medLDM}
\alias{medLDM}
\title{medLMD: A package for simulating landscape dynamics in Mediterranean regions}
\description{
The medLDM package provides a set of functions to simulate landscape-scale processes 
that influence the composition and the arrangment of the landscape such as wildfires and 
land-cover changes, as well as the structure and composition of forest stands such as 
fire and forest management and vegetation dynamics.
}
\section{Author(s)}{

\bold{Maintainer}: Núria Aquilué \email{nuria.aquilue@ctfc.cat}  

\bold{Authors}: Núria Aquilué, Lluís Brotons, Andrea Duane, Quim Canelles
}

