% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.fire.r
\name{post.fire}
\alias{post.fire}
\title{Post-fire regeneration}
\usage{
post.fire(land, clim, params)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records and land-cover types in rows}

\item{clim}{A data frame with minimum temperature (in ºC), maximum temperature (in ºC), 
and precipitation (in mm) per location}

\item{params}{A list of default parameters generated by the function \code{default.params()} or}
}
\value{
A vector with the new tree species or land-cover in the burnt locations the current time step
}
\description{
Determines the tree species or land-cover type after the impact of fire
}
\examples{
data(landscape)
data(clim)
params = default.params()
post.fire(landscape, clim, params)

}
