% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.sqi.r
\name{sdm.sqi}
\alias{sdm.sqi}
\title{Compute SDM and SQI for forest species}
\usage{
sdm.sqi(land, orography, clim)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records and land-cover types in rows}

\item{orography}{A data frame with the orographic characteristics of the study area}

\item{clim}{A data frame with default minimum temperature (in ºC), maximum temperature (in ºC), 
and precipitation (in mm) per location}
}
\value{
A data frame with the binary SDM (0 or 1) and the categorical SQI (1-low, 2-high, 3-optimal) for
each location in the study area identifed by \code{cell.id}
}
\description{
Computes the species distribution model (SDM) and the site quality index (SQI) of forest stands
}
\examples{
data(landscape)
data(orography)
data(clim)
x = sdm.sqi(landscape, orography, clim)

}
