% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{wildfires}
\alias{wildfires}
\title{Wildfires}
\format{
A data frame with 3210899 rows and 11 variables:
\describe{
  \item{cell.id}{Unique grid cell indentificator}
  \item{y10}{Locations burnt in 2010}
  \item{y11}{Locations burnt in 2011}
  \item{y12}{Locations burnt in 2012}
  \item{y13}{Locations burnt in 2013}
  \item{y14}{Locations burnt in 2014}
  \item{y15}{Locations burnt in 2015}
  \item{y16}{Locations burnt in 2016}
  \item{y17}{Locations burnt in 2017}
  \item{y18}{Locations burnt in 2018}
  \item{y19}{Locations burnt in 2019}
}
}
\usage{
wildfires
}
\description{
Observed wildfires in Catalonia from 2010 to 2019
}
\keyword{datasets}
