% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afforestation.r
\name{afforestation}
\alias{afforestation}
\title{Afforestation}
\usage{
afforestation(land, coord, orography, clim, params)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records and land-cover types in rows}

\item{coord}{A data frame with yhe oordinates x and y in UTM31N-ETRS89 of the 1 ha cells in the study area}

\item{orography}{A data frame with the orographic characteristics of the study area}

\item{clim}{A data frame with minimum temperature (in ºC), maximum temperature (in ºC), and precipitation (in mm) per location}

\item{params}{A list of model parameters, default ones are generated by the function \code{default.params()}}
}
\value{
A vector with the new tree species in colonized shurb locations
}
\description{
Determines the shrub locations colonized by tree species
}
\examples{
data(landscape)
data(coord)
data(orography)
data(clim)
params = default.params()
afforestation(landscape, coord, orography, clim, params)

}
