% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.establish.r
\name{cohort.establish}
\alias{cohort.establish}
\title{Cohort establishment}
\usage{
cohort.establish(land, coord, orography, sdm, params)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records and land-cover types in rows}

\item{coord}{A data frame with yhe oordinates x and y in UTM31N-ETRS89 of the 1 ha cells in the study area}

\item{orography}{A data frame with the orographic characteristics of the study area}

\item{sdm}{A data frame with the binary species distribution models of the tree species in the study area}

\item{params}{A list of default parameters generated by the function \code{default.params()} or}

\item{clim}{A data frame with minimum temperature (in ºC), maximum temperature (in ºC), and precipitation (in mm) per location}
}
\value{
A vector with the new tree species in after drought-induced mortality
}
\description{
Determines the tree species after forest dieback
}
\examples{
data(landscape)
data(coord)
data(orography)
data(clim)
data(sdm)
kill.cells = drought(landscape, clim, 10, 1)
landscape$tsdist[landscape$cell.id \%in\% killed.cells] = 0
landscape$typdist[landscape$cell.id \%in\% killed.cells] = "drght"
cohort.establish(landscape, coord, orography, clim, sdm)

}
