% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encroachment.r
\name{encroachment}
\alias{encroachment}
\title{Encroachment}
\usage{
encroachment(land, coord, orography)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records and land-cover types in rows}

\item{coord}{A data frame with yhe oordinates x and y in UTM31N-ETRS89 of the 1 ha cells in the study area}

\item{orography}{A data frame with the orographic characteristics of the study area}

\item{params}{A list of default parameters generated by the function \code{default.params()} or}
}
\value{
A vector with the \code{cell.id} of the colonized grassland areas
}
\description{
Determines the grassland locations colonized by shrubs
}
\examples{
data(landscape)
data(coord)
data(orography)
encroachment(landscape, coord, orography)

}
