% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.mgmt.r
\name{forest.mgmt}
\alias{forest.mgmt}
\title{Forest management}
\usage{
forest.mgmt(
  land,
  clim,
  harvest,
  demand.sawlog = 0,
  demand.wood = 0,
  policy = "BAU"
)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records and land-cover types in rows}

\item{clim}{A data frame with minimum temperature (in ºC), maximum temperature (in ºC), 
and precipitation (in mm) per location}

\item{harvest}{A \code{harvest} data frame with landscape variables driving the allocation of sylvicultural prescriptions}

\item{policy}{A string indicating which regional policy is adopted to allocate the timber demands}

\item{sawlog.demand}{Sawlog volume to be harvested (in m3)}

\item{wood.demand}{Wood for energy volume to be harvested (in m3)}
}
\value{
Data frames still to be polished (under development)!
}
\description{
Spatially allocation of sawlog and wood for energy demand according to a  land-cover transitions
}
\examples{
data(landscape)
data(clim)
data(orography)
data(harvest)
land = landscape \%>\% mutate(typcut = NA, tscut  = NA, tburnt = NA)
land = land \%>\% left_join(sdm.sqi(land, orography, clim), by="cell.id")
forest.mgmt(land, clim, harvest, 100, 80)

}
