% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.r
\name{interface}
\alias{interface}
\title{Land interfaces}
\usage{
interface(land)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records and land-cover types in rows}
}
\value{
A vector with the code interface of all locations in the study area: 
1 - urban, 2 - cropland, 3 - natural areas, 4 - other, 
5 - urban & cropland, 6 - urban & natural areas, and 7 - cropland & natural areas
}
\description{
Determines the land interface of each location
}
\examples{
data(landscape)
x = interface(landscape)

}
