% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{land.cover.changes}
\alias{land.cover.changes}
\title{Land-cover changes}
\format{
A data frame with as many rows as grid cells in the study area and 2 variables:
\describe{
  \item{cell.id}{Unique grid cell indentificator}
  \item{code}{Code made by merging the 2-digits code of the land-cover type in 1993 
  to the 2-digits code of the land-cover code in 2017. The 20-categories legend is as \code{landscape$spp}}
}
}
\usage{
land.cover.changes
}
\description{
Codification of the land-cover changes occurred between 1993 and 2017 based on the 
20-categories legend of the Land Cover Map of Catalonia
}
\keyword{datasets}
